Modificações no retorno quando a categoria não existir

+++ back-end/src/test/java/com/eventmanager/pachanga/factory/CategoriaFactoryTest.java
@@ -44,7 +44,7 @@ class CategoriaFactoryTest {
 		
 		CategoriaTO categoriaTO  = CategoriaFactory.getCategoriaTO(categoria);
 		
-		assertEquals( categoriaTO.getCodCategoria(), categoria.getCodCategoria());
+		assertEquals( categoriaTO.getCodCategoria().intValue(), categoria.getCodCategoria());
 		assertEquals( categoriaTO.getNomeCategoria(), categoria.getNomeCategoria());
 	}
